tests/wpt/tests/shared-storage/same-origin-add-module-credentials-same-origin.tentative.https.sub.html

<!doctype html>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/common/utils.js"></script>
<script src="/shared-storage/resources/util.js"></script>
<script src="/fenced-frame/resources/utils.js"></script>

<body>
<script>
'use strict';

promise_test(async () => {
  const ancestor_key = token();
  document.cookie = "key0=value0";
  const helper_url = `/shared-storage/resources/credentials-test-helper.py`;

  await sharedStorage.worklet.addModule(
    helper_url + `?action=store-cookie&token=${ancestor_key}`,
    { credentials: "same-origin" });

  const request_cookie_fetch_response =
    await fetch(helper_url + `?action=get-cookie&token=${ancestor_key}`);

  const request_cookie_text = await request_cookie_fetch_response.text();

  assert_equals(request_cookie_text, "key0=value0");
}, 'addModule() with same-origin module script and credentials "same-origin"');

</script>
</body>
